refactor: split oversized workbench modules
